So much better than primegen.py it's not funny.
C++ is so damn fast Python with multiprocessing can't beat C++ on a single thread. Although some of that might have to do with primegen.py not even being written with speed in mind like this was but still. 5 seconds vs 0.1 seconds is still a huge win no matter what in my book.